1.00_β30
MotionSupporter_1.00_β30 miumiu
・スムージング
　・不要キー削除処理を復活させた時にcdefの指定になってて、pythonから読めなくなっていたメソッドをcpdefに修正
・モーフブレンド
　・増減値の型がintになってて正確な値が受渡されてなかったのをfloatに修正
・足FKtoIK
　・足IKの回転初期値を常にクリアするよう修正
　　（前キーの値を受け継いでしまっていた）
・多段分割/統合
　・元ボーンを入力した際の補間で、移動と回転のキーを推奨多段ボーン名がある場合にはそちらを優先して設定するよう処理追加

+++ src/module/MOptions.pyx
@@ -157,7 +157,7 @@ cdef class MLegFKtoIKOptions:
 class MBlendOptions():
 
     def __init__(self, str version_name, int logging_level, PmxModel model, list eye_list, list eyebrow_list, list lip_list, list other_list, \
-                 float min_value, float max_value, int inc_value):
+                 float min_value, float max_value, float inc_value):
         self.version_name = version_name
         self.logging_level = logging_level
         self.model = model
